backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.17
sqlalchemy==2.0.48
psycopg[binary]==3.3.3
pydantic-settings==2.6.1
Pillow==10.4.0
torch==2.5.1
torchvision==0.20.1
opencv-python==4.10.0.84
websockets==15.0.1
numpy==2.1.3
librosa==0.10.2.post1
soundfile==0.13.1
joblib==1.4.2
pandas==2.2.3
scikit-learn==1.6.1
requests==2.32.3
beautifulsoup4==4.12.3
IPy>=1.1,<2
tld>=0.13,<1
xpinyin>=0.7.7,<1
